Update .travis.yml
I've previously submitted the same change but it was lost in conflict
resolution.

There is no need to add 10gen repository or install mongodb or JDK in
the travis-ci.org CI env. All VM images have mongodb and one JDK.

All you have to do to be future proof is to make sure it is started
(it is currently started on boot but we will change this in the future).