Update README.rst
Some instructions were out of date. I had to install libxslt1-dev and build-essential to get it working. I have updated the document but having to fork code for doing this is cumbersome. So I have created a Wiki page instead so that people can contribute documentation without having to fork the code.